feat: Animate waybar gpuinfo and memory module (#869)
* feat: Add animated icon for gpuinfo waybar module

Animate the thermometer icon in the gpuinfo module depending on the
current GPU temperature.

The temperature and utilization icons in tooltip are also animated.
Also restored the temperature dependent emoji in the tooltip and some
other minor changes.

* feat: Add animated icon for memory waybar module

Make the memory icon in the waybar module animated. The memory meter
rises and falls with memory usage.